+++ src/css/custom.css
@@ -71,24 +71,24 @@ navpath {
     font-weight: bold; 
 }
 
-html[data-iframe="true"] .navbar, 
-html[data-iframe="true"] header, 
-html[data-iframe="true"] footer, 
-html[data-iframe="true"] .theme-doc-sidebar-container,
-html[data-iframe="true"] .theme-doc-toc-desktop, 
-html[data-iframe="true"] .theme-doc-toc-mobile,
-html[data-iframe="true"] nav.pagination-nav, 
-html[data-iframe="true"] div[role="complementary"],
-html[data-iframe="true"] aside { 
+div[data-iframe="true"] .navbar, 
+div[data-iframe="true"] header, 
+div[data-iframe="true"] footer, 
+div[data-iframe="true"] .theme-doc-sidebar-container,
+div[data-iframe="true"] .theme-doc-toc-desktop, 
+div[data-iframe="true"] .theme-doc-toc-mobile,
+div[data-iframe="true"] nav.pagination-nav, 
+div[data-iframe="true"] div[role="complementary"],
+div[data-iframe="true"] aside { 
   display: none !important;
 }
 
-html[data-iframe="true"] main[class^="docMainContainer_"] {
+div[data-iframe="true"] main[class^="docMainContainer_"] {
   max-width: 100% !important;
   width: 100% !important;
 }
 
-html[data-iframe="true"] main[class^="docMainContainer_"] > div {
+div[data-iframe="true"] main[class^="docMainContainer_"] > div {
   padding-top: 1rem !important;
 }
